Modularize app into src package; move resources
Restructure the project: replace the large monolithic main.py with a modular src/ package and add a new main_tkinter.py entrypoint. Added src modules (constants, settings, utils), a ui package (main_window, dialogs, theme, widgets), and workers (download_worker, scrape_worker) to separate concerns and improve maintainability. Moved all static assets (icons, audio, favicon) into a resources/ directory and updated build_exe.bat to include the resources folder and use resources/main.ico. Updated settings.json accordingly.
