+++ resources/cursor-text.png
