+++ resources/gear-fill.png
