+++ src/constants.py
@@ -1,0 +1,9 @@
+from pathlib import Path
+
+APP_TITLE = "MyInstants Downloader and Player"
+APP_SETTINGS_FILE = Path("settings.json")
+DEFAULT_DOWNLOAD_DIR = Path("downloads")
+DOWNLOAD_CHUNK_SIZE = 4096
+
+DEFAULT_BASE_URL = "https://www.myinstants.com"
+DEFAULT_REGION = "us"
